services/requirements.txt
pymongo
pypdf
langchain
langchain_community
langchain-text-splitters
sentence-transformers
python-dotenv
langchain_huggingface